Refactor code

+++ 2GD15N_Kinoo_Lucas_Unity_Game/GMUnityGame/Assets/LucasKinoo/GMUnityGame/Scripts/Gameplay/DamageBehaviour.cs
@@ -8,9 +8,6 @@ public class DamageBehaviour : MonoBehaviour
     {
         PlayerCharacter player = other.GetComponent<PlayerCharacter>();
 
-        if (player != null)
-        {
-            player.Kill();
-        }
+        if (player != null) player.Kill();
     }
 }
